klist "fixing" + kfrees + proc improvements

+++ kernel/libk/aex/klist.h
@@ -21,5 +21,5 @@ void  klist_init_raw(struct klist* klist, klist_entry_t* first_entry);
 bool  klist_set(struct klist* klist, size_t index, void* ptr);
 void* klist_get(struct klist* klist, size_t index);
 
-size_t klist_first(struct klist* klist);
-void*  klist_iter(struct klist* klist, klist_entry_t** entry);
+int64_t klist_first(struct klist* klist);
+void*   klist_iter(struct klist* klist, klist_entry_t** entry);
